Added Two Functions and a Resize Listener
Both of the functions are going to look for a click on the transfer buttons for their table. When clicked it will transfer the selected row from one table to another. For the resize listener it will automatically redraw the data tables to ensure they are the appropriate size for the webpage.